requirements.txt
# Core dependencies
flask==2.3.2
langdetect==1.0.9
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.6

# Database
# Note: sqlite3 is part of Python standard library, not a PyPI package
# sqlite3==2.6.0

# Media processing
whisper==1.1.10
openai-whisper==20231117
pytube==15.0.0
yt-dlp==2023.11.16
moviepy==1.0.3
ffmpeg-python==0.2.0

# YouTube API and transcription
google-api-python-client==2.108.0
youtube-transcript-api==0.6.1
isodate==0.6.1

# AI and NLP
openai==0.27.8
transformers==4.31.0
torch==2.0.1
numpy==1.24.3
sentence-transformers==2.2.2
scikit-learn==1.3.0
google-generativeai==0.3.1
absl-py==1.4.0

# News APIs
newsapi-python==0.2.7
feedparser==6.0.10
beautifulsoup4==4.12.2

# Kafka integration
confluent-kafka==2.2.0
avro==1.11.1
kafka-python==2.0.2

# OCR for image processing
pytesseract==0.3.10
Pillow==10.0.0

# Web scraping
lxml==4.9.3

# Logging and utilities
tqdm==4.66.1
termcolor==2.3.0 